feat: add ModelSelector demo with slots and variants
